Updating to grad accum of fp32 for BF16 ZeRO1 case.

+++ tests/unit/runtime/test_ds_initialize.py
@@ -172,7 +172,7 @@ class TestOptimizerImplementation(DistributedTest):
         # ZeRO 1 Wrapper
         is_supported[('zero1', 'fp16', None)] = True
         is_supported[('zero1', 'fp16', 'fp16')] = True
-        is_supported[('zero1', 'bf16', 'bf16')] = True
+        is_supported[('zero1', 'bf16', 'fp32')] = True
         is_supported[('zero1', 'fp32', None)] = True
         is_supported[('zero1', 'fp32', 'fp32')] = True
         # ZeRO 2 Wrapper
